feat(marketing): add adoption playbook and boost landing conversion path
